requirements.txt
flask
finnhub-python
websocket
flask-redis
flask-sqlalchemy
pyfcm
requests
pymysql
PyJWT
bcrypt
pytest
pytest-flask